Rename and fix files location
