+++ lib/application/client/search/bloc/client_search_bloc.dart
@@ -2,11 +2,11 @@ import 'dart:async';
 
 import 'package:appointment/application/client/search/client_search_status.dart';
 import 'package:appointment/domain/client/client_entity.dart';
-import 'package:appointment/infrastructure/client/client_filters.dart';
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/core/select_filter.dart';
-import 'package:appointment/infrastructure/core/page_service.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
+import 'package:appointment/infrastructure/drift/client/client_filters.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/core/select_filter.dart';
+import 'package:appointment/infrastructure/drift/core/page_service.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
 import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
