+++ lib/infrastructure/drift/client/client_filters.dart
@@ -1,6 +1,6 @@
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/core/select_filter.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/core/select_filter.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
 import 'package:drift/drift.dart';
 
 class ClientNameFilter implements SelectFilter<ClientModels, ClientModel> {
