+++ lib/infrastructure/drift/client/client_table.dart
@@ -1,6 +1,6 @@
 // This file is used for code generation only, it should not be tested
 // coverage:ignore-file
-import 'package:appointment/infrastructure/common/type_converters.dart';
+import 'package:appointment/infrastructure/drift/common/type_converters.dart';
 import 'package:drift/drift.dart';
 
 class ClientModels extends Table {
