+++ lib/infrastructure/drift/drift_db.dart
@@ -1,10 +1,10 @@
 import 'package:appointment/domain/client/client_values.dart';
 import 'package:appointment/domain/common/common_values.dart';
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/common/type_converters.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/common/type_converters.dart';
 import 'package:drift/drift.dart';
 
-part 'db.g.dart';
+part 'drift_db.g.dart';
 
 @DriftDatabase(tables: [ClientModels])
 class DriftDb extends _$DriftDb {
