+++ lib/infrastructure/drift/executor/executor_provider.dart
@@ -1,0 +1,3 @@
+export 'unsupported_executor.dart'
+    if (dart.library.js) 'web_executor.dart'
+    if (dart.library.ffi) 'native_executor.dart';
