--- lib/infrastructure/drift/executor_provider.dart
@@ -1,3 +1,0 @@
-export 'unsupported_db.dart'
-    if (dart.library.js) 'web_db_factory.dart'
-    if (dart.library.ffi) 'native_db_factory.dart';
