+++ lib/presentation/config/di_config.dart
@@ -11,15 +11,15 @@ import 'package:appointment/application/register/bloc/register_bloc.dart';
 import 'package:appointment/application/register/register_validator.dart';
 import 'package:appointment/domain/client/client_entity.dart';
 import 'package:appointment/domain/core/repository.dart';
-import 'package:appointment/infrastructure/client/client_converter.dart';
-import 'package:appointment/infrastructure/client/client_dao.dart';
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/core/dao.dart';
-import 'package:appointment/infrastructure/core/entity_model_converter.dart';
-import 'package:appointment/infrastructure/core/page_service.dart';
-import 'package:appointment/infrastructure/core/drift_repository.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
-import 'package:appointment/infrastructure/drift/executor_provider.dart'
+import 'package:appointment/infrastructure/drift/client/client_converter.dart';
+import 'package:appointment/infrastructure/drift/client/client_dao.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/core/dao.dart';
+import 'package:appointment/infrastructure/drift/core/entity_model_converter.dart';
+import 'package:appointment/infrastructure/drift/core/page_service.dart';
+import 'package:appointment/infrastructure/drift/core/drift_repository.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
+import 'package:appointment/infrastructure/drift/executor/executor_provider.dart'
     as executor_provider;
 import 'package:drift/drift.dart';
 import 'package:get_it/get_it.dart';
