+++ test/application/client/search/bloc/client_search_bloc_test.dart
@@ -1,10 +1,10 @@
 import 'package:appointment/application/client/search/bloc/client_search_bloc.dart';
 import 'package:appointment/application/client/search/client_search_status.dart';
 import 'package:appointment/domain/client/client_entity.dart';
-import 'package:appointment/infrastructure/client/client_filters.dart';
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/core/page_service.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
+import 'package:appointment/infrastructure/drift/client/client_filters.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/core/page_service.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
 import 'package:bloc_test/bloc_test.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
