+++ test/infrastructure/client/client_converter_test.dart
@@ -1,8 +1,8 @@
 import 'package:appointment/domain/client/client_entity.dart';
 import 'package:appointment/domain/client/client_values.dart';
 import 'package:appointment/domain/common/common_values.dart';
-import 'package:appointment/infrastructure/client/client_converter.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
+import 'package:appointment/infrastructure/drift/client/client_converter.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
