+++ test/infrastructure/core/page_service_test.mocks.dart
@@ -1,13 +1,14 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in appointment/test/infrastructure/client/page_service_test.dart.
+// in appointment/test/infrastructure/core/page_service_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
 import 'package:appointment/domain/common/common_values.dart' as _i5;
-import 'package:appointment/infrastructure/core/dao.dart' as _i3;
-import 'package:appointment/infrastructure/core/select_filter.dart' as _i6;
+import 'package:appointment/infrastructure/drift/core/dao.dart' as _i3;
+import 'package:appointment/infrastructure/drift/core/select_filter.dart'
+    as _i6;
 import 'package:drift/drift.dart' as _i1;
 import 'package:mockito/mockito.dart' as _i2;
