+++ test/presentation/client/details/widgets/client_details_form_test.dart
@@ -6,7 +6,7 @@ import 'package:appointment/application/edit/bloc/edit_bloc.dart';
 import 'package:appointment/domain/client/client_entity.dart';
 import 'package:appointment/domain/client/client_values.dart';
 import 'package:appointment/domain/common/common_values.dart';
-import 'package:appointment/infrastructure/client/client_dao.dart';
+import 'package:appointment/infrastructure/drift/client/client_dao.dart';
 import 'package:appointment/presentation/app_ointment.dart';
 import 'package:appointment/presentation/client/common/widgets/client_name_form_field.dart';
 import 'package:appointment/presentation/client/details/widgets/client_details_name_input.dart';
