+++ test/presentation/client/register/client_register_page_test.dart
@@ -2,9 +2,9 @@ import 'package:appointment/application/client/bloc/client_bloc.dart';
 import 'package:appointment/application/client/register/client_register_validator.dart';
 import 'package:appointment/application/register/bloc/register_bloc.dart';
 import 'package:appointment/domain/client/client_entity.dart';
-import 'package:appointment/infrastructure/client/client_table.dart';
-import 'package:appointment/infrastructure/core/drift_repository.dart';
-import 'package:appointment/infrastructure/drift/db.dart';
+import 'package:appointment/infrastructure/drift/client/client_table.dart';
+import 'package:appointment/infrastructure/drift/core/drift_repository.dart';
+import 'package:appointment/infrastructure/drift/drift_db.dart';
 import 'package:appointment/presentation/client/register/client_register_page.dart';
 import 'package:appointment/presentation/client/register/widgets/client_register_form.dart';
 import 'package:flutter/material.dart';
